Move get_domain_ids and output_file_lists to postprocessing_common
