+++ src/fsipy/automatedPostprocessing/postprocessing_fenics/create_hdf5.py
@@ -4,12 +4,12 @@
 
 import numpy as np
 import h5py
-import re
 from pathlib import Path
 import json
 import logging
 
 from fsipy.automatedPostprocessing.postprocessing_fenics import postprocessing_fenics_common
+from fsipy.automatedPostprocessing.postprocessing_common import get_domain_ids, output_file_lists
 from dolfin import Mesh, HDF5File, VectorFunctionSpace, Function, MPI, parameters
 
 
@@ -172,87 +172,6 @@ def create_hdf5(visualization_path, mesh_path, save_time_step, stride, start_tim
 
     logging.info("--- Finished reading solutions")
     logging.info(f"--- Saved u.h5 and d.h5 in {visualization_separate_domain_folder.absolute()}")
-
-
-def get_domain_ids(mesh_path, fluid_domain_id, solid_domain_id):
-    """
-    Given a mesh file, this function returns the IDs of the fluid and solid domains
-
-    Args:
-        mesh_path (Path): Path to the mesh file that contains the fluid and solid domains
-        fluid_domain_id (int): ID of the fluid domain
-        solid_domain_id (int): ID of the solid domain
-
-    Returns:
-        fluid_ids (list): List of IDs of the fluid domain
-        solid_ids (list): List of IDs of the solid domain
-        all_ids (list): List of IDs of the whole mesh
-    """
-    with h5py.File(mesh_path) as vector_data:
-        domains = vector_data['domains/values'][:]
-        topology = vector_data['domains/topology'][:, :]
-
-        if isinstance(fluid_domain_id, list):
-            id_fluid = np.where((domains == fluid_domain_id[0]) | (domains == fluid_domain_id[1]))
-        else:
-            id_fluid = np.where(domains == fluid_domain_id)
-
-        if isinstance(solid_domain_id, list):
-            id_solid = np.where((domains == solid_domain_id[0]) | (domains == solid_domain_id[1]))
-        else:
-            id_solid = np.where(domains == solid_domain_id)
-
-        wall_topology = topology[id_solid, :]
-        fluid_topology = topology[id_fluid, :]
-
-        # Get topology of fluid, solid and whole mesh
-        solid_ids = np.unique(wall_topology)
-        fluid_ids = np.unique(fluid_topology)
-        all_ids = np.unique(topology)
-
-    return fluid_ids, solid_ids, all_ids
-
-
-def output_file_lists(xdmf_file):
-    """
-    If the simulation has been restarted, the output is stored in multiple files and may not have even temporal spacing
-    This loop determines the file names from the xdmf output file
-
-    Args:
-        xdmf_file (Path): Path to xdmf file
-
-    Returns:
-        h5file_name_list (list): List of names of h5 files associated with each timestep
-        timevalue_list (list): List of time values in xdmf file
-        index_list (list): List of indices of each timestp in the corresponding h5 file
-    """
-
-    file1 = open(xdmf_file, 'r')
-    Lines = file1.readlines()
-    h5file_name_list = []
-    timevalue_list = []
-    index_list = []
-
-    # This loop goes through the xdmf output file and gets the time value (timevalue_list), associated
-    # with .h5 file (h5file_name_list) and index of each timestep in the corresponding h5 file (index_list)
-    for line in Lines:
-        if '<Time Value' in line:
-            time_pattern = '<Time Value="(.+?)"'
-            time_str = re.findall(time_pattern, line)
-            time = float(time_str[0])
-            timevalue_list.append(time)
-
-        elif 'VisualisationVector' in line:
-            h5_pattern = '"HDF">(.+?):/'
-            h5_str = re.findall(h5_pattern, line)
-            h5file_name_list.append(h5_str[0])
-
-            index_pattern = "VisualisationVector/(.+?)</DataItem>"
-            index_str = re.findall(index_pattern, line)
-            index = int(index_str[0])
-            index_list.append(index)
-
-    return h5file_name_list, timevalue_list, index_list
 
 
 def main() -> None:
